requirements.txt
streamlit
ultralytics
opencv-python
Pillow
pandas
numpy
